fix: Print Qty should be Int

+++ erpnext/accounts/doctype/sales_invoice/sales_invoice.js
@@ -248,7 +248,7 @@ erpnext.accounts.SalesInvoiceController = erpnext.selling.SellingController.exte
 				in_list_view: 1,
 				label: __('Contents Qty')
 			}, {
-				fieldtype: 'Float',
+				fieldtype: 'Int',
 				fieldname: "print_qty",
 				in_list_view: 1,
 				label: __('Print Qty')
